bitrock-airways-10 Servizio per acquisto Ticket - seconda parte

+++ src/main/java/it/bitrock/bitrockairways/repository/TicketRepository.java
@@ -19,7 +19,7 @@ public interface TicketRepository extends JpaRepository<Ticket, Integer> {
     List<Ticket> getTicketsByFlight_Route(Airport airport);
 
     @Query("select t from Ticket t WHERE t.flight.id = :flightId")
-    List<Ticket> getTicketsByFlightId(Integer flightId);
+    List<Ticket> getTicketsByFlightId(Long flightId);
 
     @Query("select t from Ticket t join t.customer where t.customer = :customer and t.dateBought < current_date")
     List<Ticket> getTicketsByCustomerBeforeNow(Customer customer);
